+++ src/App.tsx
@@ -8,13 +8,13 @@ import {
   Routes,
   Route,
 } from "react-router-dom";
-import Shell from "src/Shell";
-import Login from "src/Login";
-import Callback from "src/Callback";
-import Dashboard from "src/Dashboard";
-import Logout from "src/Logout";
-import Profile from "src/Profile";
-import Submissions from "src/Submissions";
+import Shell from "src/screens/Shell";
+import Login from "src/screens/Login";
+import Callback from "src/components/Callback";
+import Logout from "src/components/Logout";
+import Dashboard from "src/screens/Dashboard";
+import Profile from "src/screens/Profile";
+import Submissions from "src/screens/Submissions";
 
 import { ProtectedRoute } from "src/components/ProtectedRoute";
 
@@ -23,7 +23,13 @@ function App() {
     <Router>
       <Routes>
         {/* Protected Routes */}
-        <Route path="home/*" element={<Shell />}>
+        <Route
+          path="home/*"
+          element={
+            <ProtectedRoute>
+              <Shell />
+            </ProtectedRoute>
+          }>
           <Route path="dashboard" element={<Dashboard />}></Route>
           <Route path="profile" element={<Profile />}></Route>
           <Route path="submissions" element={<Submissions />}></Route>
